Add rendering and separate classes in segmentation package
rendering - will contain classes to render labels, to start with it has
a class that draws the outlines of a segmented image on a grayscale.

separate - will contain classes to separate touching objects, to start
with it has a 'separate_touching' method that uses binary operations,
distance map and watershed.