Add a second version of separate
This is called 'separate_touching2'.  It uses a hybrid of the scikit
strategy and the clij strategy for pre-processing to watershed.  Like
the clij strategy it applies a blur to the original image to detect
local max to be used for seeds.  Then a second (usually smaller) blur
is applied top the original image and used to create a distance map,
seed points, distance map, and Otsu of the original are then fed to the
scikitimage watershed function.