Add filter_with_df and classify_with_df
These functions are helpers to be used when one creates a dataframe of
region properties from a labeled image.  These functions create an image
from the labeled image and dataframe, that only contains the objects in
the dataframe.  It is useful for creating afiltered label image after  applying
pandas operations to filter the dataframe.

Classify_with_df expects a class column in the dataframe.  It is useful
when trying to visualize classes from the labeled image.