Add a bit of nuance regarding the perfect estimation
If we estimate the object perfectly the 'image' will only be equal to
the true image within the window.  Brings up the question of whether to
get better 'out of window' reconstruction should we extend even further?
Probably no practical use, but something to think about, and make sure
I'm not still having a slight error.