Misc. updates to notebooks
This commit adds some misc. updates to the notebooks.  Currently this
entire codebase is at the early stages of development.  Thus minor
changes are made to the notebooks fairly often to experiment with
different images or settings.

At this stage the notebooks still may be useful as examples, but look
carefully for any mistakes or flaws as they are still changing often.

Eventually the notebooks may become more stable and a testing framework
added.