Add/tweak deep learning examples
Continue the beginning of the deep learning example set.  Restore.ipynb
shows intensity restoration using a very trivial shallow conv-net, this
is expected to fail because the receptive field is to small to restore
large objects and blur functions.  Then we show a related example that
uses Care (which in turn uses unets), and has a much larger receptive
field thus provides much better results.