Start framework to keep track of models
We need a way to keep track of all the models we are using.  Models can
be 'deep learning' models, Point Spread Functions or possibly other.

The idea is we put information about the model, specifically which
instrument or instruments it is created/calibrated/trained for.

For the case of a PSF it would be one system that the PSF was generated
for (if a theoretical PSF) or the parameters of the system the bead
image was taken on (if extracted PSF).

For the case of a deep learning model it would be the system parameters
of the instrument the training image were taken on.  For deep learning
their can potentially by a one (model) to many (system parameters).
That the DL model may have been trained on images from several systems.