Improve dl_helper
add the following parameters to collect_training_data

training_multiple: forces training size to be a multiple of
traing_multiple.  For example if training_multiple is 16, the size will
be cropped to the nearest multiple of 16 (some DL networks need the size
to be a multiple of n, where n is often 16).

patch_size: Crops data to patch_size if data is bigger than patch_size
(uses random crop)

Also add shuffle function, which shuffles an input (X) and ground truth
(Y) set, but keeps them consistent with each other.