Add add_small_to_large_2d
This function could be much DRYer (ie just modify add_small_to_large to
work on ND).  That is on the todo list.  However for now these are
handy utilities to help create phantom and label images so we will
accept the warts and just accept that 'perfect is the enemy of the good'
or 'dry code is the enemy of 'it works let's go home'' (kidding)