Add make_patch_source_divide2 utility
This is a hackish little utility to take an image and simply divide it
in two.  One half can be used to create patches for a DL system.  The
other half can be kept 'unseen' for testing.  This was inspired by
recent questions where the poster only has one image example and we want
to keep some of the data hidden from the DL.