Also keep a copy of fft_filtering under the imagesc directory
This isn't very DRY at all.  However some notebooks are very good
examples of a general image processing workflow (in this case FFT
filtering) so it is useful to but them under a categorical folder.
However it is also useful to keep them underneath the imagesc folder
organized by date of question.