Move example to imagesc folder
Sometimes I've been calling imagesc example 'listserv'.  The reason is
historical (before imagesc there was an imagej listserv, so I got in the
habit of naming examples for imagej listserv questions as 'listserv
...'.  I'm now trying to organize imagesc examples in an imagesc folder
sorted by date.  So over the next year hoping to go back to examples
I've done for imagesc questions and put them into the imagesc folder.

To start with this little example shows how to explore how a 3D PSF will
look in 2D in the presence of defocus and aberrations.