Using plt to create figures is now optional
A 'use_plt' parameter has been added.  If True figures will be generated
using the plt interface (thus invoking the GUI).  If false figures will
be created with the 'Figure' constructor.

We need to be able to bypass plt when generating figure in worker
threads or other situations where we don't want the figure to be
displayed.  However keeping the option of using plt is handy because it
makes the figures automatically show in notebooks and scripts.