Tweak x,y,z order
I couldn't get meshgrid to work right with z,y,x order even though it
should be trivial.  For some reason the z seemed to be put in the middle
coordinate... When I have time I need to look into this further.

For now just make sure when using the coordinates returned from meshgrid
that the order (x,y,z) is used properly.