Add albumentations, matplotlib, scipy and tiffile as dependencies
I have not used any dependendencies for tnia-python so far because it
was intended as a utility and example project and each utility and/or
example had a different set of dependencies.  However since the
utilities for creating sets of augmented patches are pretty useful, and
being used in other projects like napari-easy-augment-batch I am going
to start handling dependencies better for tnia-python.  More to come but
to start with add pypi dependencies for albumentations, matplotlib,
scipy, and tifffile.