Document plottypes
Plottypes are 'imshow' and 'hist'.  'imshow' plot with matplotlib.pyplot
imshow, and hist generates a histogram.  This is handy because we can
easily generate multi=plots of an image and corresponding histogram by
putting the image in twice in the ims array, and setting the plottypes
array to ['imshow', 'hist']