Add many sams and cyto3 shared memory example
This example allows us to run cellposesam, and cellposecyto3 in the same
notebook and then view the results in the same napari viewer.  We also
take a look at the microsam and mobilesam results.

Cellposesam and cellposecyto3 have to be in different environments, so
we use shared memory and subprocess to call them.